src/requirements.txt
Flask
python-plexapi
requests
